// src/types/gitStats.ts
export interface ContributorNode {
  id: string;
  name: string;
}

export interface ContributorEdge {
  source: string;
  target: string;
  weight: number;
}